2.1.(document).ready() & Ordering
When we use this javascripts block thing:

 72 lines  app/Resources/views/lift/index.html.twig
{% block javascripts %}
    {{ parent() }}
    <script>
        $('.js-delete-rep-log').on('click', function() {
            console.log('todo delete!');
        });
    </script>
{% endblock %}
It adds our new JavaScript code right after the main script tags in the base layout:

 97 lines  app/Resources/views/base.html.twig
<html lang="en">
<body>
{% block javascripts %}
    <script src="https://code.jquery.com/jquery-3.1.1.min.js" integrity="sha256-hVVnYaiADRTO2PzUGmuLJr8BLUSjGIZsDYGmIJLv2b8=" crossorigin="anonymous"></script>
    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
{% endblock %}
</body>
</html>
View the HTML source and scroll to the bottom see that in action. Yep, jQuery and then our stuff.

Our JavaScript lives at the bottom of the page for a reason: performance. Unless you add an async attribute, when your browser sees a script tag, it stops, waits while that file is downloaded, executes it, and then continues.

But not everyone agrees that putting JS in the footer is the best thing since Chuck Norris. After all, if your page is heavily dependent on JS, your user might see a blank page for a second before your JavaScript has the chance to execute and put cool stuff there, like a photo of Chuck Norris.