2.3.(document).ready() & Ordering{Running JavaScript Before the DOM}
Running JavaScript Before the DOM

Our browser executes JavaScript as soon as it sees it... which might be before some or all of the page has actually loaded. Our code is looking for all elements with the js-delete-rep-log class. Well, at this point, none of the HTML body has loaded yet, so it finds exactly zero elements.

This is the reason why you probably already always use the famous $(document).ready() block. Move our code inside of it, and refresh again

Yes!

Very simply, jQuery calls your $(document).ready() function once the DOM has fully loaded. But it's nothing fancy: it's approximately equal to putting your JavaScript code at the absolute bottom of the page. It's nice because it makes our code portable: it will work no matter where it lives.

We could even take the script tag, delete it from the block, and put it right in the middle of the page

Now in the HTML, the external script tags are still on top, but our JavaScript lives right, smack in the middle of the page. And when we refresh, it still works super well.