3.2.All about Event Bubbling
I'm feeling so good about our first click listener, let's add another! When I click anywhere on a row, I also want to log a message.

Back in the template, give the entire table a js class so we can select it. How about js-rep-log-table.

Down below, find that and look inside for the tbody tr elements. Then, .on('click') add a function that prints some fascinating text: console.log('row clicked')

Beautiful! Refresh and click the row. No surprises: we see "row clicked". But check this out: click the delete link. Hot diggity - two log messages! Of course it would do this! I clicked the delete link, but the delete link is inside of the row. Both things got clicked!

All about Event Bubbling
Welcome to event bubbling, an important concept in JavaScript that's just boring enough that you've probably avoided reading articles about it in the past. Let's make it awesome.

Here it goes: when we click, we cause a click event. Now technically, when I click the delete icon, the element that I'm actually clicking is the span that holds the icon. Cool! So, your browser goes to that span element and says:

  Top of the morning! I'd like to trigger a click event on you!

Then, if there are any listener functions attached on click, those are called. Next, your browser goes up one level to the anchor and says:

  Ahoy Matey! I'd like to trigger a click event on you!

And the same thing happens again: if there are any click listener functions attached to that element, those are executed. This includes our listener function. From here, it just keeps going: bubbling all the way up the tree: to the td, the tr, tbody, table, and eventually, to the <body> tag itself.

And that is why we see "todo delete" first: the event bubbling process notifies the link element and then bubles up and notifies the tr.