4.2.The Event Argument & stopPropagation{The Event (e) Listener Argument}
The Event (e) Listener Argument

Back up a step. Whenever a listener function is called, your browser passes it an event argument, commonly just named e.

This e variable is packed with information and some functions. The most important is e.preventDefault(). Another is e.stopPropagation().

It turns out that when you return false from a listener function, it is equivalent to calling e.preventDefault() and e.stopPropagation(). To prove it, remove the return false and refresh.

Yep, same behavior: no # sign, but still no "row clicked" when we click the delete icon.