5.2.The DOM Element Object{Pro Tip: Using console.dir()}
Pro Tip: Using console.dir()

And there's another way you can see the methods and properties on this object. Go back and change your console.log() to console.dir()

Now refresh. Click a link and check this out! It still gives you some information about what the element is, but now you can expand it to find a huge list of its properties and methods. Nice! One of the properties is called className, which we will use in a second.

If you're not familiar with console.dir(), it's bananas cool. Sometimes, console.log() gives you a string representation of something. But console.dir() tries to give you a tree of what that thing actually is. It's like programmer X-Ray vision!