6.3.The Magical this Variable & currentTarget{Enter: this (versus currentTarget)}
Enter: this (versus currentTarget)

In fact, try this: console.log(e.currentTarget === this).

Refresh! And click anywhere on the delete link. It's always true.

There's a good chance that you've been using the this variable for years inside of your listener functions to find the element that was clicked. And now we know the true and dramatic story behind it! this is equivalent to e.currentTarget, the DOM Element that we originally attached our listener to.

Ultimately that means that we can say, $(this).addClass('text-danger'). That will always add the text-danger link to the anchor tag.