7.3.A Great Place to Hide Things! The data- Attributes{Adding data-weight Metadata}
Adding data-weight Metadata

If we somehow knew what the weight was for this specific row, we could read the total weight and just subtract it when it's deleted. So how can we figure out the weight for this row? Well, we could just read the HTML of the third column... but that's kinda shady. Instead, why not use another data- attribute?

On the <tr> element, add a data-weight attribute set to repLog.totalWeightLifted. Also, so that we know which th to update, add a class: js-total-weight

Let's hook this up! Before the AJAX call - that's important, we'll find out why soon - find the total weight container by saying $table.find('.js-total-weight'). Next add var newWeight set to $totalWeightContainer.html() - $row.data('weight'). Use that inside success: $totalWeightContainer.html(newWeight)

Let's give this fanciness a try. Go back refresh. 459? Hit delete, it's gone. 454.

Now, how about we get into trouble with some JavaScript objects!