8.2.Organizing with Objects! {Initializing (not Instantiating) the Object}
Initializing (not Instantiating) the Object

Back in the (document).ready(), our job is really simple: find the $table and then pass it to RepLogApp.initialize()

The cool thing about this approach is that now we have an entire object who's job is to work inside of this..

Ok, let's try this! Go back and refresh! Hit delete! Ah, it fails!

  Variable $table is not defined.

The problem is inside of handleRepLogDelete. Ah, cool, this makes total sense. Before, we had a $table variable defined above the function. That's gone