9.2."Static" Objects & the this Variable {Refactoring to More Methods! }
Refactoring to More Methods!
Since we're running out of items, let's add a few more!

Now that we have a fancy object, we can use it to get even more organized, by breaking big functions into smaller ones.

For example, we could create a new function called, updateTotalWeightLifted. Instead of figuring out the total weight lifted here and doing the update down in the success callback. We'll just call this method and have it do all that heavy lifting.

Add var totalWeight = 0

Then I'll say, this.$wrapper, which I can do because we're not in a callback function: this is our object. Then, .find to look for all tbody tr elements, and .each() to loop over them

But stop! Notice that when you use .each(), you pass it a callback function! So guess what? Inside, this is no longer our RepLogApp object, it's something different. In this case, this is the individual tr DOM Element object that we're looping over in this moment.

Inside, add up all the total weights with totalWeight += $(this).data() and read the data-weight attribute. Finally use this.$wrapper.find() to look for our js-total-weight element and set its HTML to totalWeight

Cool!

Down in handleRepLogDelete, we don't need any of this logic anymore, nor this logic. We just need to call our new function. The only gotcha is that the fadeOut() function doesn't actually remove the row from the DOM, so our new weight-totaling function would still count its weight.

Fix it by telling fadeOut() to use normal speed, pass it a function to be called when it finishes fading, and then say $row.remove() to fully remove it from the DOM

Now we can call updateTotalWeightLifted.

But check this out: we're actually inside of another callback function, which is inside of a callback function, inside of our entire function which is itself a callback! So, this is definitely not our RepLogApp object.

No worries, play it safe and use RepLogApp.updateTotalWeightLifted() instead

That's the equivalent in PHP of calling a static method by using its class name.

Ok, try it out! Refresh the page. We're at 765. Now delete a row... 657! Nice! Let's finally figure out what's really going on with the this variable... and how to make it act better!