10.2.Getting to the bottom of the this Variable{Being a Magician with this! }
Being a Magician with this!

Let's have a little fun with the weirdness of this. Create a new function - just for debugging - called whatIsThis with a single argument, a greeting. Inside, just console.log() this and our greeting

Next, at the bottom of initialize, add this.whatIsThis() and pass it hello.

Simple enough! And since we're calling this function directly - not as a callback - I would expect this to actually be what we expect: our RepLogApp object. Let's find out. Refresh! Expand the logged object. Yea, it's RepLogApp! Cool!