13.1.The window Object & Global Variables {What is this window?}
Now that we're using this fancy self-executing function, we don't have access to RepLogApp anymore.How can we fix that? Very simple. Instead of var RepLogApp, say window.RepLogApp. Back in the template, I'll delete the console.log() for Helper

And then go back and refresh. It works! No error in the console, and delete does its job!

What is this window?

So what the heck just happened? Here's the deal: when you're executing JavaScript in a browser - which for you is probably always - you always have access to a global window variable. In fact, it's even more important than that. This window variable holds all of the global variables. What I mean is: if you set a key on the window object, like RepLogApp, this becomes a global variable. That means you can reference RepLogApp from anywhere else, and this is actually referencing window.RepLogApp. More on that in a second.