13.2.The window Object & Global Variables {Passing Yourself Global Variables}
Passing Yourself Global Variables

Inside of our self-executing function, we - of course - also have access to any global variables, like window or the $ jQuery variable. But, instead of relying on these global variables, you'll often see people pass those variables into the function. It's a little weird, so let's see it.

Right now, inside of our self-executing function, we're using two global variables: window and $, for $.ajax, for example

At the bottom of the file, between the parentheses, reference the global window and jQuery variables and pass them as arguments to our function. On top, add those arguments: window and $

Now, when we reference window and $ in our code, we're no longer referencing the global objects directly, we're referencing those arguments.

Why the heck would you do this? There are two reasons, and neither are huge. First, you can alias global variables. At the bottom, we reference the jQuery global variable, which is even better than referencing $ because sometimes people setup jQuery in no conflict mode, where it does not create a $ variable. But then above, we alias this to $, meaning it's safe inside for us to use that shortcut. You probably don't have this problem, but you'll see stuff like this in third-party libraries.

Second, when you pass in a global variable as an argument, it protects you from making a really silly mistake in your code, like accidentally setting $ = null. If you do that now, it'll set $ to null only inside this function. But before, you would have overwritten that variable globally. It's yet another way that self-executing blocks help to sandbox us.