13.3.The window Object & Global Variables {Fun with window}
Fun with window

Ok, back to this mysterious window variable. Inside index.html.twig, console.log() window.

This is pretty cool, because it will show us all global variables that are available.

And Boom! This is a huge object, and includes the $ variable, jQuery, and eventually, RepLogApp.

But notice what's not here. As expected, there is no Helper.