13.5.The window Object & Global Variables {Forget var? It goes Global! }
Be Better: use strict

Back in RepLogApp, forgetting var is actually a mistake, but JavaScript is friendly, and it allows us to make mistakes. In real life, friendly and forgiving people are great friends! In programming, friendly and forgiving languages mean more bugs!

To tell JavaScript to stop being such a pushover, at the top of the RepLogApp.js file, inside quotes, say 'use strict'

 Tip

Even better! Put 'use strict' inside the self-executing function. Adding 'use strict' applies to the function its inside of and any functions inside of that (just like creating a variable with var). If you add it outside of a function (like we did), it affects the entire file. In this case, both locations are effectively identical. But, if you use a tool that concatenates your JS files into a single file, it's safer to place 'use strict' inside the self-executing function, to ensure it doesn't affect those other concatenated files!

I know, weird. This is a special JavaScript directive that tells your browser to activate a more strict parsing mode. Now, certain things that were allowed before, will cause legit errors. And sure enough, when we refresh, we get:

Uncaught reference: Helper is not defined

Sweeeet! Even PhpStorm isn't fooled anymore, it's reporting an:

Unresolved variable or type Helper

Re-add var, and life is good